+++ src/main/utils.ts
@@ -1,0 +1,42 @@
+import fs from 'node:fs';
+import os from 'node:os';
+import path from 'node:path';
+import { dialog } from 'electron';
+import log from 'electron-log';
+import type { Menubar } from 'menubar';
+
+export function takeScreenshot(mb: Menubar) {
+  const date = new Date();
+  const dateStr = date.toISOString().replace(/:/g, '-');
+
+  const capturedPicFilePath = `${os.homedir()}/${dateStr}-atlassify-screenshot.png`;
+  mb.window.capturePage().then((img) => {
+    fs.writeFile(capturedPicFilePath, img.toPNG(), () =>
+      log.info(`Screenshot saved ${capturedPicFilePath}`),
+    );
+  });
+}
+
+export function resetApp(mb: Menubar) {
+  const cancelButtonId = 0;
+  const resetButtonId = 1;
+
+  const response = dialog.showMessageBoxSync(mb.window, {
+    type: 'warning',
+    title: 'Reset Atlassify',
+    message:
+      'Are you sure you want to reset Atlassify? You will be logged out of all accounts',
+    buttons: ['Cancel', 'Reset'],
+    defaultId: cancelButtonId,
+    cancelId: cancelButtonId,
+  });
+
+  if (response === resetButtonId) {
+    mb.window.webContents.send('atlassify:reset-app');
+    mb.app.quit();
+  }
+}
+
+export function getIconPath(iconName: string) {
+  return path.resolve(__dirname, '../assets/images', iconName);
+}
